Create SI_FSM_v3.vhd
std_logic_unsigned library removed.
rising_edge() function is used in index_REG_proc (address_register).
HM_proc deleted, a when-else statement used instead.
FSM_comb changed accordingly; the state transitions are changed so that the states turn back to s1 when the safety decision is finalized, the different_index check is put at the topmost of the process since it is the first condition to be true for a state transition to occur, some if conditions are simplified according to the possible cases the condition can get (e.g. (cache_HM = '0' OR cache_HM = '1') checks are deleted since they are always true) or the state transition possibilities (e.g. condition check in s2 is now done only according to cache_HM since the process ID does not affect the decision and the condition check in s3 was simplified to one if block since only one occurrence leads to unsafe output).
reg_safe_count and reg_index_count are defined as unsigned.